+++ src/components/Navbars/AuthNavbar.js
@@ -59,6 +59,7 @@ export default function Navbar(props) {
                 <button
                   className="bg-orange-500 text-white active:bg-lightBlue-600 text-xs font-bold uppercase px-4 py-2 rounded shadow hover:shadow-lg outline-none focus:outline-none lg:mr-1 lg:mb-0 ml-3 mb-3 ease-linear transition-all duration-150"
                   type="button"
+                  onClick={() => window.location.href = "/auth"}
                 >
                   콘솔에 로그인
                 </button>