+++ src/layouts/Admin.js
@@ -11,8 +11,8 @@ import FooterAdmin from "components/Footers/FooterAdmin.js";
 // views
 
 import Dashboard from "views/admin/Dashboard.js";
-import Maps from "views/admin/Maps.js";
-import Settings from "views/admin/Settings.js";
+import CreateTunnel from "views/admin/CreateTunnel.js";
+import UserProfile from "views/admin/UserProfile.js";
 import Tables from "views/admin/Tables.js";
 import {validateTokenAxios} from '../axios/login.axios';
 
@@ -22,6 +22,7 @@ export default function Admin() {
     if (token) {
       validateTokenAxios(token).then((response) => {
         if (!response.data.snsId) {
+          localStorage.clear();
           window.location.href = '/';
         }
       })
@@ -40,8 +41,8 @@ export default function Admin() {
         <div className="px-4 md:px-10 mx-auto w-full -m-24">
           <Switch>
             <Route path="/admin/dashboard" exact component={Dashboard} />
-            <Route path="/admin/maps" exact component={Maps} />
-            <Route path="/admin/settings" exact component={Settings} />
+            <Route path="/admin/add" exact component={CreateTunnel} />
+            <Route path="/admin/settings" exact component={UserProfile} />
             <Route path="/admin/tables" exact component={Tables} />
             <Redirect from="/admin" to="/admin/dashboard" />
           </Switch>