+++ src/views/admin/CreateTunnel.js
@@ -1,0 +1,144 @@
+import React from "react";
+
+export default function CreateTunnel() {
+  return (
+    <>
+      <div className="flex flex-wrap">
+        <div className="w-full lg:w-8/12 px-4">
+          <div className="relative flex flex-col min-w-0 break-words w-full mb-6 shadow-lg rounded-lg bg-blueGray-100 border-0">
+            <div className="rounded-t bg-white mb-0 px-6 py-6">
+              <div className="text-center flex justify-between">
+                <h6 className="text-blueGray-700 text-xl font-bold">새 터널 만들기</h6>
+              </div>
+            </div>
+            <div className="flex-auto px-4 lg:px-10 py-10 pt-0">
+              <form>
+                <h6 className="text-blueGray-400 text-sm mt-3 mb-6 font-bold uppercase">
+                  기본 정보
+                </h6>
+                <div className="flex flex-wrap">
+                  <div className="w-full lg:w-12/12 px-4">
+                    <div className="relative w-full mb-3">
+                      <label
+                        className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
+                        htmlFor="grid-password"
+                      >
+                        서버 이름
+                      </label>
+                      <input
+                        type="text"
+                        className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
+                        placeholder="관리에 사용할 고유한 이름 입니다."
+                      />
+                    </div>
+                  </div>
+                  <div className="w-full lg:w-6/12 px-4">
+                    <div className="relative w-full mb-3">
+                      <label
+                        className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
+                        htmlFor="grid-password"
+                      >
+                        서버 타입
+                      </label>
+                      <select
+                        name="region"
+                        className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
+                        defaultValue="Lucky"
+                      >
+                        <option value="minecraft_java">마인크래프트 자바 에디션</option>
+                      </select>
+                    </div>
+                  </div>
+                </div>
+
+                <hr className="mt-6 border-b-1 border-blueGray-300" />
+                <h6 className="text-blueGray-400 text-sm mt-3 mb-6 font-bold uppercase">
+                  터널 접속 정보
+                </h6>
+                <div className="flex flex-wrap">
+                  <div className="w-full lg:w-6/12 px-4">
+                    <div className="relative w-full mb-3">
+                      <label
+                        className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
+                        htmlFor="grid-password"
+                      >
+                        서브 도메인 (최소 4자, 최대 30지)
+                      </label>
+                      <input
+                        type="text"
+                        className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
+                        placeholder="서버 접속에 필요한 주소"
+                      />
+                    </div>
+                  </div>
+                  <div className="w-full lg:w-6/12 px-4">
+                    <div className="relative w-full mb-3">
+                      <label
+                        className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
+                        htmlFor="grid-password"
+                      >
+                        루트 도메인
+                      </label>
+                      <select
+                        name="rootDomain"
+                        className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
+                        defaultValue="Lucky"
+                      >
+                        <option value="blbt.app">blbt.app</option>
+                        <option value="waterflake.co">waterflake.co</option>
+                      </select>
+                    </div>
+                  </div>
+                  <div className="w-full lg:w-12/12 px-4">
+                    <div className="relative w-full mb-3">
+                      <label
+                        className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
+                        htmlFor="grid-password"
+                      >
+                        게임 서버 접속 주소
+                      </label>
+                      <input
+                        type="text"
+                        readOnly={true}
+                        value="minecraft.blbt.app"
+                        className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
+                      />
+                    </div>
+                  </div>
+                  <div className="w-full lg:w-6/12 px-4">
+                    <div className="relative w-full mb-3">
+                      <label
+                        className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
+                        htmlFor="grid-password"
+                      >
+                        리전
+                      </label>
+                      <select
+                        name="region"
+                        className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
+                        defaultValue="Lucky"
+                      >
+                        <option value="KR-SUWON-1">KR-SUWON-1</option>
+                      </select>
+                    </div>
+                  </div>
+                  <div className="w-full lg:w-6/12 px-4"></div>
+                </div>
+                <hr className="mt-6 border-b-1 border-blueGray-300" />
+                <div className="w-full lg:w-6/12 px-4">
+                  <br />
+                  <button
+                    className="bg-emerald-500 active:bg-lightBlue-600 uppercase text-white font-bold hover:shadow-md shadow px-4 py-2 rounded outline-none focus:outline-none sm:mr-2 mb-1 ease-linear transition-all duration-150"
+                    type="button"
+                  >
+                    <i className="fas fa-check mr-2 text-sm"></i>{" "}터널 생성
+                  </button>
+                </div>
+              </form>
+            </div>
+          </div>
+        </div>
+      </div>
+    </>
+  );
+}