+++ src/views/admin/UserProfile.js
@@ -1,0 +1,120 @@
+import React from "react";
+
+// components
+
+import CardSettings from "components/Cards/CardSettings.js";
+import CardProfile from "components/Cards/CardProfile.js";
+
+export default function UserProfile() {
+  return (
+    <>
+      <div className="flex flex-wrap">
+        <div className="w-full lg:w-8/12 px-4">
+          <div className="relative flex flex-col min-w-0 break-words w-full mb-6 shadow-lg rounded-lg bg-blueGray-100 border-0">
+            <div className="rounded-t bg-white mb-0 px-6 py-6">
+              <div className="text-center flex justify-between">
+                <h6 className="text-blueGray-700 text-xl font-bold">내 계정</h6>
+              </div>
+            </div>
+            <div className="flex-auto px-4 lg:px-10 py-10 pt-0">
+              <h6 className="text-blueGray-400 text-sm mt-3 mb-6 font-bold uppercase">
+                기본 정보
+              </h6>
+              <div className="flex flex-wrap">
+                <div className="w-full lg:w-6/12 px-4">
+                  <div className="relative w-full mb-3">
+                    <label
+                      className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
+                      htmlFor="grid-password"
+                    >
+                      사용자 이름
+                    </label>
+                    <input
+                      type="text"
+                      readOnly={true}
+                      className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
+                      defaultValue="볼봇"
+                    />
+                  </div>
+                </div>
+                <div className="w-full lg:w-6/12 px-4">
+                  <div className="relative w-full mb-3">
+                    <label
+                      className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
+                      htmlFor="grid-password"
+                    >
+                      이메일 주소
+                    </label>
+                    <input
+                      type="email"
+                      readOnly={true}
+                      className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
+                      defaultValue="jesse@example.com"
+                    />
+                  </div>
+                </div>
+                <div className="w-full lg:w-6/12 px-4">
+                  <div className="relative w-full mb-3">
+                    <label
+                      className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
+                      htmlFor="grid-password"
+                    >
+                      가입일
+                    </label>
+                    <input
+                      type="text"
+                      readOnly={true}
+                      className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
+                      defaultValue="2023년 04월 17일"
+                    />
+                  </div>
+                </div>
+              </div>
+
+              <hr className="mt-6 border-b-1 border-blueGray-300" />
+
+              <h6 className="text-blueGray-400 text-sm mt-3 mb-6 font-bold uppercase">
+                고급 정보
+              </h6>
+              <div className="flex flex-wrap">
+                <div className="w-full lg:w-12/12 px-4">
+                  <div className="relative w-full mb-3">
+                    <label
+                      className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
+                      htmlFor="grid-password"
+                    >
+                      고유 ID
+                    </label>
+                    <input
+                      type="text"
+                      readOnly={true}
+                      className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
+                      defaultValue="snsid here"
+                    />
+                  </div>
+                </div>
+                <div className="w-full lg:w-4/12 px-4">
+                  <div className="relative w-full mb-3">
+                    <label
+                      className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
+                      htmlFor="grid-password"
+
+                    >
+                      인증 제공자
+                    </label>
+                    <input
+                      type="email"
+                      readOnly={true}
+                      className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
+                      defaultValue="KAKAO"
+                    />
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </>
+  );
+}