+++ CHANGELOG.md
@@ -38,7 +38,7 @@
 ### Added components
 ### Deleted dependencies
 - `@tailwindcss/custom-forms`
-- `react-google-maps` (replaced by simple Google Maps API)
+- `react-google-maps` (replaced by simple Google CreateTunnel API)
 - `@types/googlemaps` (dependencies of `react-google-maps`)
 - `@types/markerclustererplus` (dependencies of `react-google-maps`)
 - `@types/react` (dependencies of `react-google-maps`)